Add basic logging features
In `main.go`, logger can be a global variable,
so any function could be using it without need
to pass it as parameter.

In `explorer.go`, logger is required to log
requests operations, since it's not happening
in the main thread.
In case nil is passed, it will create a default
`log.Logger`.

Refs #4